pyproject.toml
[project]
name = "paralympics"
version = "1.0.0"
description = "Paralympics testing activities"
dependencies = [
    "flask",
    "Flask-SQLAlchemy",
    "Flask-Marshmallow",
    "marshmallow-sqlalchemy",
    "pandas",
    "selenium",
    "pytest"
]

# https://setuptools.pypa.io/en/latest/userguide/pyproject_config.html
[build-system]
requires = [
    "setuptools",
    "setuptools-scm",
]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["paralympics"]  # list of folders that contain the packages (["."] by default)
include = ["paralympics*"]  # package names should match these glob patterns (["*"] by default)
exclude = ["tests*"]  # exclude packages matching these glob patterns (empty by default)
namespaces = false  # to disable scanning PEP 420 namespaces (true by default)

# https://docs.pytest.org/en/stable/reference/customize.html#pyproject-toml
[tool.pytest.ini_options]
addopts = "--verbose"
testpaths = ["tests"]
xfail_strict = true
filterwarnings = [
    "error",
    # pytest's own futurewarnings
    "ignore::pytest.PytestExperimentalApiWarning",
]

# https://flask.palletsprojects.com/en/2.3.x/tutorial/tests/#running-the-tests
[tool.coverage.run]
branch = true
source = ["paralympics"]
